Move image classes
